SAM/clip_crop/app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
pillow==8.3.2
git+https://github.com/openai/CLIP.git
seaborn==0.9.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
urllib3==1.26.6
